pyproject.toml
[project]
name = "proy-ssdd-ii"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.8",
    "confluent-kafka>=2.14.0",
    "contextily>=1.7.0",
    "geopandas>=1.1.3",
    "holidays>=0.93",
    "ipykernel>=7.2.0",
    "jinja2>=3.1.6",
    "logging>=0.4.9.6",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "orjson>=3.11.8",
    "pandas>=3.0.1",
    "pathlib>=1.0.1",
    "pyarrow>=23.0.1",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "shapely>=2.1.2",
    "ujson>=5.12.0",
]
